Updated fibonacci types
Changed from ints to unsigned longs to support larger numbers (still works)

+++ algorithms/include/algo.hpp
@@ -5,9 +5,9 @@
 
 
 
-long fibonacci(const int n)
+unsigned long fibonacci(const unsigned long n)
 {
-    int a = 0, b = 1, c = 0;
+    unsigned long a = 0, b = 1, c = 0;
 
     if (n == 0 || n == 1)
     {
@@ -15,7 +15,7 @@ long fibonacci(const int n)
     } 
     else
     {
-        for (int i = 2; i <= n; i++) {
+        for (unsigned long i = 2; i <= n; i++) {
             c = a + b;
             a = b;
             b = c;
